Replace some entity parameter and variables in the serviceImpls package with DTO - 1

+++ src/main/java/com/umutdoruk/hrms/service/serviceImpls/CityServiceImpl.java
@@ -1,50 +1,67 @@
 package com.umutdoruk.hrms.service.serviceImpls;
 
+import com.umutdoruk.hrms.DTO.request.CityRequest;
+import com.umutdoruk.hrms.DTO.response.CityResponse;
 import com.umutdoruk.hrms.entities.City;
 import com.umutdoruk.hrms.exception.NotFoundException;
 import com.umutdoruk.hrms.repository.CityRepository;
 import com.umutdoruk.hrms.service.services.CityService;
+import com.umutdoruk.hrms.service.services.JobAdvertisementService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class CityServiceImpl implements CityService {
 
     private final CityRepository cityRepository;
+    private final JobAdvertisementService jobAdvertisementService;
 
     @Autowired
-    public CityServiceImpl(CityRepository cityRepository) {
+    public CityServiceImpl(CityRepository cityRepository, JobAdvertisementService jobAdvertisementService) {
         this.cityRepository = cityRepository;
+        this.jobAdvertisementService = jobAdvertisementService;
     }
 
     @Override
-    public void create(City city) {
+    public void create(CityRequest cityRequest) {
+
+        if (cityRequest == null)
+            throw new NotFoundException("No City record found to add");
+
+        City city = new City();
+        city.setCityName(cityRequest.getCityName());
+        city.setJobAdvertisement(jobAdvertisementService.findById(cityRequest.getJobAdvertisementId()));
+
         cityRepository.save(city);
     }
 
     @Override
-    public void update(String cityName) {
-        City city = cityRepository.findByCityName(cityName);
-        city.setCityName(cityName);
+    public void update(CityRequest cityRequest, Long cityId) {
+
+        City city = cityRepository.findById(cityId)
+                        .orElseThrow(()-> new NotFoundException("No City with this Id in Repository"));
+
+        if (cityRequest == null)
+            throw new NotFoundException("No City record found to update");
+
+        city.setCityName(cityRequest.getCityName());
+        city.setJobAdvertisement(jobAdvertisementService.findById(cityRequest.getJobAdvertisementId()));
+
         cityRepository.save(city);
     }
 
     @Override
     public void delete(Long id) {
+        if (!(cityRepository.existsById(id)))
+            throw new NotFoundException("No City found to delete");
         cityRepository.deleteById(id);
     }
 
     @Override
-    public List<City> getAll() {
-        return cityRepository.findAll();
-    }
-
-    @Override
-    public City getById(Long id) {
-        return cityRepository.findById(id)
+    public CityResponse getById(Long id) {
+         City city = cityRepository.findById(id)
                 .orElseThrow(()-> new NotFoundException("City is not found"));
+        return CityResponse.of(city);
     }
 
 }
